added endpoints documenttion using swagger UI
