+++ package.json
@@ -2,16 +2,10 @@
   "name": "my-brand",
   "version": "1.0.0",
   "description": "Andela capstone project build my first API of my personal website",
-  "main": "server.mjs",
-
+  "main": "server.js",
   "scripts": {
-    "start": "npm run build && node ./build/server.mjs",
-    "start:dev": "nodemon --exec babel-node ./server.mjs",
-    "clean": "rm -rf build && mkdir build",
-    "build-babel": "babel -d ./build ./server.mjs -s",
-    "build": "npm run clean && npm run build-babel",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint mjs"
+    "start:dev": "nodemon --exec babel-node ./server.js",
+    "build-babel": "babel -d ./build ./server.js -s"
   },
   "author": "Rachel",
   "license": "ISC",
@@ -22,7 +16,9 @@
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^6.2.4"
+    "mongoose": "^6.2.4",
+    "swagger-jsdoc": "^6.1.0",
+    "swagger-ui-express": "^4.3.0"
   },
   "devDependencies": {
     "@babel/node": "^7.16.8",
@@ -31,5 +27,6 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/runtime": "^7.17.2",
     "eslint": "^8.10.0"
-  }
+  },
+  "type":"module"
 }
